Set Select Action to Registration Cntr

+++ app/controllers/users/registrations_controller.rb
@@ -6,6 +6,8 @@ class Users::RegistrationsController < Devise::RegistrationsController
 
 
   def select  ##登録方法の選択ページ
+    redirect_to root_path, alert: "ログインしています。" if user_signed_in?
+    @auth_text = "で登録する"
     session.delete(:"devise.sns_auth") if session["devise.sns_auth"]
   end
